clean up plugin file stucture and support multiple plugin setups
