--- apps/teskooano/src/config/componentRegistry.ts
@@ -1,8 +1,0 @@
-import type { ComponentRegistryConfig } from "@teskooano/ui-plugin";
-
-// This file now only contains application-specific components.
-// The core components are loaded via the vite.config.ts
-export const componentConfig: ComponentRegistryConfig = {
-  // REMOVED "teskooano-simulation-controls"
-  // REMOVED "teskooano-system-controls"
-};
