+++ apps/teskooano/src/core/components/button/README.md
@@ -1,0 +1,22 @@
+# Core Button Component (`<teskooano-button>`)
+
+A standard button component for the Teskooano UI.
+
+## Usage
+
+```html
+<teskooano-button>Click Me</teskooano-button>
+
+<teskooano-button variant="primary">Primary Action</teskooano-button>
+
+<teskooano-button variant="secondary" disabled>Disabled</teskooano-button>
+```
+
+## Attributes
+
+- `variant`: (Optional) Controls the button's appearance. Values: `default` (or omitted), `primary`, `secondary`, `danger`.
+- `disabled`: (Optional) Standard boolean attribute to disable the button.
+
+## Events
+
+- Standard `click` event.
