--- apps/teskooano/src/core/config/componentRegistry.ts
@@ -1,3 +1,0 @@
-import type { ComponentRegistryConfig } from "@teskooano/ui-plugin";
-
-export const componentConfig: ComponentRegistryConfig = {};
