+++ apps/teskooano/src/core/config/pluginRegistry.ts
@@ -1,0 +1,13 @@
+import type { PluginRegistryConfig } from "@teskooano/ui-plugin";
+
+export const pluginConfig: PluginRegistryConfig = {
+  "core-button": { path: "../core/components/button" },
+  "core-modal": { path: "../core/components/modal" },
+  "core-card": { path: "../core/components/card" },
+  "core-output": { path: "../core/components/output" },
+  "core-select": { path: "../core/components/select" },
+  "core-slider": { path: "../core/components/slider" },
+  "core-tooltip": { path: "../core/components/tooltip" },
+  "core-tour": { path: "../core/interface/tour-controller" },
+  "core-engine-toolbar": { path: "../core/interface/engine-toolbar" },
+};
